+++ watcher/test/watcher-enterprise.test.js
@@ -25,7 +25,9 @@ const {
   DATASETORDERS_COLLECTION,
   WORKERPOOLORDERS_COLLECTION,
   REQUESTORDERS_COLLECTION,
+  fastForwardToLastBlock,
 } = require('./test-utils');
+const { init: ethereumInit } = require('../src/loaders/ethereum');
 
 jest.setTimeout(120000);
 
@@ -1876,8 +1878,7 @@ describe('Recover on start enterprise specific', () => {
 describe('Replay Past enterprise specific', () => {
   beforeAll(async () => {
     await dropDB(chainId);
-    await start({ syncWatcher: false, replayer: false });
-    await stop();
+    await ethereumInit();
   });
 
   beforeEach(async () => {
@@ -1887,6 +1888,7 @@ describe('Replay Past enterprise specific', () => {
       WORKERPOOLORDERS_COLLECTION,
       REQUESTORDERS_COLLECTION,
     ]);
+    await fastForwardToLastBlock(chainId, rpc);
     jest.clearAllMocks();
   });
 
@@ -1982,6 +1984,7 @@ describe('Replay Past enterprise specific', () => {
     await revokeKYC(userAddress);
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     expect(socketEmitSpy).toHaveBeenCalledTimes(0);
+    await fastForwardToLastBlock(chainId, rpc);
     await replayPastOnly({ nbConfirmation: 0 });
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     const [
@@ -2162,6 +2165,7 @@ describe('Replay Past enterprise specific', () => {
     await revokeKYC(userAddress);
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     expect(socketEmitSpy).toHaveBeenCalledTimes(0);
+    await fastForwardToLastBlock(chainId, rpc);
     await replayPastOnly({ nbConfirmation: 0 });
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     const [
@@ -2372,6 +2376,7 @@ describe('Replay Past enterprise specific', () => {
     await revokeKYC(userAddress);
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     expect(socketEmitSpy).toHaveBeenCalledTimes(0);
+    await fastForwardToLastBlock(chainId, rpc);
     await replayPastOnly({ nbConfirmation: 0 });
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     const [
@@ -2581,6 +2586,7 @@ describe('Replay Past enterprise specific', () => {
     await revokeKYC(userAddress);
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     expect(socketEmitSpy).toHaveBeenCalledTimes(0);
+    await fastForwardToLastBlock(chainId, rpc);
     await replayPastOnly({ nbConfirmation: 0 });
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     const [
@@ -2771,6 +2777,7 @@ describe('Replay Past enterprise specific', () => {
     await grantKYC(userAddress);
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     expect(socketEmitSpy).toHaveBeenCalledTimes(0);
+    await fastForwardToLastBlock(chainId, rpc);
     await replayPastOnly({ nbConfirmation: 0 });
     await sleep(PROCESS_TRIGGERED_EVENT_TIMEOUT);
     const [